Lasso selection cleanup
Lasso selection has been refactored/cleaned up to ensure better consistency
through vispy examples and further optimization have been brought.
